requirements.txt
allure-pytest==2.6.2
allure-python-commons==2.6.2
atomicwrites==1.3.0
attrs==19.1.0
enum34==1.1.6
lxml==4.3.3
more-itertools==7.0.0
namedlist==1.7
pluggy==0.9.0
py==1.8.0
pytest==4.4.0`
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
